+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
   - Added test for bug [#166](https://sourceforge.net/p/simplehtmldom/bugs/166/)
 - Added unit test for character sets UTF-8, CP1251 and CP1252 (#142)
 - Added support for meta charset to parse_charset
+- Added detection for CP1251 to parse_charset, using iconv
 - Added LICENSE file (MIT) to the project root
 ### Changed
 - Changed manual from custom implementation to MkDocs (https://www.mkdocs.org/)